Created module for complete scenarios and base class for criteria
Introduction of a Scenario module (Scenarios/), which contains complete
scenarios (incl. vehicle setup, spawn points, behavior, etc.).
Currently, the only scenario in there, is the "follow a leading vehicle"
scenario.

In addition, a base class for all evaluation criteria was introduced to
simplify adding more criteria.
